move the get_config function
this function isn't a decorator so it's a bit confusing that it was in the decorators.py file
since it only exists to cache part of the `/config` api response, let's put it next to that
api definition
